backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
elevenlabs==1.9.0
httpx==0.27.2
supabase==2.7.4
python-dotenv==1.0.1
python-multipart==0.0.12
websockets==13.1